+++ test/common/services/ReportExporter.test.ts
@@ -1,20 +1,21 @@
 import { Mock } from 'ts-mockery';
 import { CsvJson } from 'csvjson';
-import { VisitsExporter } from '../../../src/visits/services/VisitsExporter';
+import { ReportExporter } from '../../../src/common/services/ReportExporter';
 import { NormalizedVisit } from '../../../src/visits/types';
 import { windowMock } from '../../mocks/WindowMock';
+import { ExportableShortUrl } from '../../../src/short-urls/data';
 
-describe('VisitsExporter', () => {
+describe('ReportExporter', () => {
   const toCSV = jest.fn();
   const csvToJsonMock = Mock.of<CsvJson>({ toCSV });
-  let exporter: VisitsExporter;
+  let exporter: ReportExporter;
 
   beforeEach(jest.clearAllMocks);
   beforeEach(() => {
     (global as any).Blob = class Blob {}; // eslint-disable-line @typescript-eslint/no-extraneous-class
     (global as any).URL = { createObjectURL: () => '' };
 
-    exporter = new VisitsExporter(windowMock, csvToJsonMock);
+    exporter = new ReportExporter(windowMock, csvToJsonMock);
   });
 
   describe('exportVisits', () => {
@@ -35,11 +36,36 @@ describe('VisitsExporter', () => {
 
       exporter.exportVisits('my_visits.csv', visits);
 
-      expect(toCSV).toHaveBeenCalledWith(visits, { headers: 'key' });
+      expect(toCSV).toHaveBeenCalledWith(visits, { headers: 'key', wrap: true });
     });
 
     it('skips execution when list of visits is empty', () => {
       exporter.exportVisits('my_visits.csv', []);
+
+      expect(toCSV).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('exportShortUrls', () => {
+    it('parses provided short URLs to CSV', () => {
+      const shortUrls: ExportableShortUrl[] = [
+        {
+          shortUrl: 'shortUrl',
+          visits: 10,
+          title: '',
+          createdAt: '',
+          longUrl: '',
+          tags: '',
+        },
+      ];
+
+      exporter.exportShortUrls(shortUrls);
+
+      expect(toCSV).toHaveBeenCalledWith(shortUrls, { headers: 'key', wrap: true });
+    });
+
+    it('skips execution when list of visits is empty', () => {
+      exporter.exportShortUrls([]);
 
       expect(toCSV).not.toHaveBeenCalled();
     });