+++ test/short-urls/ShortUrlsList.test.tsx
@@ -6,7 +6,6 @@ import shortUrlsListCreator from '../../src/short-urls/ShortUrlsList';
 import { ShortUrlsOrderableFields, ShortUrl, ShortUrlsOrder } from '../../src/short-urls/data';
 import { MercureBoundProps } from '../../src/mercure/helpers/boundToMercureHub';
 import { ShortUrlsList as ShortUrlsListModel } from '../../src/short-urls/reducers/shortUrlsList';
-import { OrderingDropdown } from '../../src/utils/OrderingDropdown';
 import Paginator from '../../src/short-urls/Paginator';
 import { ReachableServer } from '../../src/servers/data';
 import { Settings } from '../../src/settings/reducers/settings';
@@ -34,6 +33,7 @@ describe('<ShortUrlsList />', () => {
           tags: [ 'test tag' ],
         }),
       ],
+      pagination: {},
     },
   });
   const ShortUrlsList = shortUrlsListCreator(ShortUrlsTable, ShortUrlsFilteringBar);
@@ -58,7 +58,6 @@ describe('<ShortUrlsList />', () => {
 
   it('wraps expected components', () => {
     expect(wrapper.find(ShortUrlsTable)).toHaveLength(1);
-    expect(wrapper.find(OrderingDropdown)).toHaveLength(1);
     expect(wrapper.find(Paginator)).toHaveLength(1);
     expect(wrapper.find(ShortUrlsFilteringBar)).toHaveLength(1);
   });
@@ -84,39 +83,26 @@ describe('<ShortUrlsList />', () => {
 
     expect(renderIcon('visits').props.currentOrder).toEqual({});
 
-    wrapper.find(OrderingDropdown).simulate('change', 'visits');
+    (wrapper.find(ShortUrlsFilteringBar).prop('handleOrderBy') as Function)('visits');
     expect(renderIcon('visits').props.currentOrder).toEqual({ field: 'visits' });
 
-    wrapper.find(OrderingDropdown).simulate('change', 'visits', 'ASC');
+    (wrapper.find(ShortUrlsFilteringBar).prop('handleOrderBy') as Function)('visits', 'ASC');
     expect(renderIcon('visits').props.currentOrder).toEqual({ field: 'visits', dir: 'ASC' });
   });
 
   it('handles order through table', () => {
     const orderByColumn: (field: ShortUrlsOrderableFields) => Function = wrapper.find(ShortUrlsTable).prop('orderByColumn');
 
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({});
+    expect(wrapper.find(ShortUrlsFilteringBar).prop('order')).toEqual({});
 
     orderByColumn('visits')();
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field: 'visits', dir: 'ASC' });
+    expect(wrapper.find(ShortUrlsFilteringBar).prop('order')).toEqual({ field: 'visits', dir: 'ASC' });
 
     orderByColumn('title')();
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field: 'title', dir: 'ASC' });
+    expect(wrapper.find(ShortUrlsFilteringBar).prop('order')).toEqual({ field: 'title', dir: 'ASC' });
 
     orderByColumn('shortCode')();
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field: 'shortCode', dir: 'ASC' });
-  });
-
-  it('handles order through dropdown', () => {
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({});
-
-    wrapper.find(OrderingDropdown).simulate('change', 'visits', 'ASC');
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field: 'visits', dir: 'ASC' });
-
-    wrapper.find(OrderingDropdown).simulate('change', 'shortCode', 'DESC');
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field: 'shortCode', dir: 'DESC' });
-
-    wrapper.find(OrderingDropdown).simulate('change', undefined, undefined);
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({});
+    expect(wrapper.find(ShortUrlsFilteringBar).prop('order')).toEqual({ field: 'shortCode', dir: 'ASC' });
   });
 
   it.each([
@@ -126,6 +112,6 @@ describe('<ShortUrlsList />', () => {
   ])('has expected initial ordering', (initialOrderBy, field, dir) => {
     const wrapper = createWrapper(initialOrderBy);
 
-    expect(wrapper.find(OrderingDropdown).prop('order')).toEqual({ field, dir });
+    expect(wrapper.find(ShortUrlsFilteringBar).prop('order')).toEqual({ field, dir });
   });
 });